==> 02_data_sampling.ipynb <==
**Enhancing Customer Retention through Machine Learning**: <br> Segmentation, Churn Prediction, and Product Recommendation – A Case Study on Flipkart Data.
<br><br>
Phase: **Data Preprocessing** <br>
By: Gia-My Nguyen <br>
Last updated (dd/mm/yyyy): 26/09/2025
<br><br>

**Completed:**
- Sampling: proportional ratios by city and 1% of full sales dataset.
- Create master dataset from sales + products (sampled).
<br><br>

**Tasks Remaining**
- No



import pandas as pd

# Get Sales data

df_sales = pd.read_csv('data/cleaned/sales_cleaned.csv')

# Get Products data

df_products = pd.read_csv('data/cleaned/products_cleaned.csv')

# Data Sampling

# Take 1% from each city 

df_sampled_1pct = df_sales.groupby("city_name", group_keys=False).apply(lambda x: x.sample(frac=0.01, random_state=42))
# df_sampled_1pct.to_csv("data/sampling/sales_1pct.csv", index=False)

#

print("1% Sample - City distribution:")
print(df_sampled_1pct["city_name"].value_counts())

# Sampled Master Data

df_master_1pct = pd.merge(
    df_sampled_1pct, df_products,
    on="product_id",
    how="left" 
)

# to csv

df_master_1pct.to_csv("data/sampling/master_1pct.csv", index=False)

df_master_1pct.head(5)

print(df_master_1pct[df_master_1pct['product_type'] == 'Unknown'].shape[0])

df_master_1pct[df_master_1pct['product_type'] == 'Unknown'].head(5)
